Factored out glob parsing.
